Single process calls in step and propagate
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@5620 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/kernel/core.cc
@@ -172,15 +172,25 @@ namespace Gecode {
   bool
   Space::propagate(unsigned long int& pn) {
     assert(!failed());
-    // Process modified variables (from initializing or from commit)
-    process();
-    Propagator* p;
-    PropModEvent pme_mask = 0;
-    while (pool_get(p)) {
-      pn++;
+    // To initialize pointer
+    PropModEvent  pme_dummy = 0;
+    // Point to propagator modevent of last executed propagators
+    PropModEvent* pme_prop = &pme_dummy;
+    // Xor mask for last executed propagator
+    PropModEvent  pme_mask = 0;
+    while (true) {
+      // Process modified variables (in first iteration: from commit or setup)
+      process();
+      // Reset propagator modevent
+      *pme_prop ^= pme_mask;
+      Propagator* p;
+      if (!pool_get(p))
+        return true;
+      pn++; pme_prop = &p->u.pme;
       switch (p->propagate(this)) {
       case ES_FAILED:
-        fail(); return false;
+        fail(); 
+        return false;
       case ES_FIX:
         // Put propagator in idle queue
         p->unlink(); a_actors.head(p);
@@ -210,7 +220,8 @@ namespace Gecode {
          * This is quite dirty: the propagator is already reused,
          * even though its pme might be used in processing. However,
          * this is safe: only the pme is used, and the pme will not be
-         * overwritten by reusing the memory.
+         * overwritten by reusing the memory (unless memory is requested,
+         * which does not happen during processing).
          */
         reuse(p,p->u.size);
         // Prevent that propagator gets rescheduled (turn on all events)
@@ -234,9 +245,8 @@ namespace Gecode {
       default:
         GECODE_NEVER;
       }
-      process();
-      p->u.pme ^= pme_mask;
     }
+    GECODE_NEVER;
     return true;
   }
 
@@ -268,70 +278,83 @@ namespace Gecode {
     if (failed())
       return ES_FAILED;
 
-    process();
-
-    Propagator* p;
-    if (!pool_get(p))
-      return ES_STABLE;
-
-    ExecStatus es = p->propagate(this);
-
-    switch (es) {
-    case ES_FAILED:
-      fail();
-      break;
-    case ES_FIX:
-      {
-        // Prevent that propagator gets rescheduled (turn on all events)
-        p->u.pme = AllVarConf::pme_assigned;
-        process();
-        p->u.pme = 0;
+    // Point to propagator after one process iteration
+    Propagator* p = NULL;
+    // Xor mask for last executed propagator
+    PropModEvent pme_mask = 0;
+    // ExecStatus of last executed propagator
+    ExecStatus es = ES_STABLE;
+    while (true) {
+      // Process modified variables (in first iteration: from commit or setup)
+      process();
+      // Reset propagator modevent
+      if (p != NULL) {
+        p->u.pme ^= pme_mask;
+        return es;
+      }
+      if (!pool_get(p))
+        return ES_STABLE;
+      es = p->propagate(this);
+      switch (es) {
+      case ES_FAILED:
+        fail(); 
+        return ES_FAILED;
+      case ES_FIX:
         // Put propagator in idle queue
         p->unlink(); a_actors.head(p);
-      }
-      break;
-    case ES_NOFIX:
-      {
+        // Prevent that propagator gets rescheduled (turn on all events)
+        p->u.pme = AllVarConf::pme_assigned;
+        // The mask resets all events after processing (idle)
+        pme_mask = AllVarConf::pme_assigned;
+        break;
+      case ES_NOFIX:
         // Propagator is currently in no queue, put into idle
         p->unlink(); a_actors.head(p);
+        // Mark propagator as idle
         p->u.pme = 0;
-        process();
-      }
-      break;
-    case __ES_SUBSUMED:
-      {
-        // Remember size
-        size_t s = p->u.size;
-        // Prevent that propagator gets rescheduled (turn on all events)
-        p->u.pme = AllVarConf::pme_assigned;
-        process();
+        // The mask leaves the pme unchanged
+        pme_mask = 0;
+        break;
+      case __ES_SUBSUMED:
+        /*
+         * Even though the propagator is already subsumed, it might
+         * still be contained in the dependency arrays of assigned
+         * variables. This is due to the fact that propagators
+         * are not unsubscribed from assigned variable implementations.
+         */
+        // Unlink propagator from queue
         p->unlink();
-        reuse(p,s);
-      }
-      break;
-    case __ES_FIX_PARTIAL:
-      {
-        // Remember the event set to be kept after processing
-        PropModEvent keep = p->u.pme;
+        /*
+         * This is quite dirty: the propagator is already reused,
+         * even though its pme might be used in processing. However,
+         * this is safe: only the pme is used, and the pme will not be
+         * overwritten by reusing the memory (unless memory is requested,
+         * which does not happen during processing).
+         */
+        reuse(p,p->u.size);
         // Prevent that propagator gets rescheduled (turn on all events)
         p->u.pme = AllVarConf::pme_assigned;
-        process();
-        p->u.pme = keep;
-        assert(p->u.pme != 0);
+        // The mask does not matter
+        break;
+      case __ES_FIX_PARTIAL:
+        // Schedule propagator with specified propagator events
         pool_put(p);
-      }
-      break;
-    case __ES_NOFIX_PARTIAL:
-      {
+        // Remember the event set to be kept after processing
+        pme_mask = AllVarConf::pme_assigned ^ p->u.pme;
+        // Prevent that propagator gets rescheduled (turn on all events)
+        p->u.pme = AllVarConf::pme_assigned;
+        break;
+      case __ES_NOFIX_PARTIAL:
         // Start from the specified propagator events
         pool_put(p);
-        process();
+        // The mask leaves the pme unchanged
+        pme_mask = 0;
+        break;
+      default:
+        GECODE_NEVER;
       }
-      break;
-    default:
-      GECODE_NEVER;
     }
-
+    GECODE_NEVER;
     return es;
   }
